requirements.txt
numpy==1.24.1
transformers==4.26.1
spacy==3.5.0
pyspellchecker==0.7.1